Made suggested change.
